changed feature_inference.py to type_inference.py and moved it to core/utils (out of tabular/utils)
