+++ deepchecks/tabular/dataset.py
@@ -22,8 +22,8 @@ from sklearn.model_selection import train_test_split
 from typing_extensions import Literal as L
 
 from deepchecks.core.errors import DatasetValidationError, DeepchecksNotSupportedError, DeepchecksValueError
-from deepchecks.tabular.utils.feature_inference import (infer_categorical_features, infer_numerical_features,
-                                                        is_categorical)
+from deepchecks.core.utils.type_inference import (infer_categorical_features, infer_numerical_features,
+                                                  is_categorical)
 from deepchecks.tabular.utils.task_type import TaskType
 from deepchecks.utils.dataframes import select_from_dataframe
 from deepchecks.utils.logger import get_logger
